requirements.txt
opencv-python
numpy
ultralytics
tensorflow
pygame
python-dotenv